+++ examples/example-04-cardinality-estimation.py
@@ -14,10 +14,10 @@ from typing import Optional
 
 import postbound as pb
 from postbound.qal import qal, transform
-from postbound import db
+from postbound import db, util
+from postbound.db import postgres
 from postbound.experiments import workloads
 from postbound.optimizer import stages, jointree, physops, planparams
-from postbound import util
 
 warnings.simplefilter("ignore")
 
@@ -74,7 +74,7 @@ class JitteringCardinalityEstimator(stages.ParameterGeneration):
 
 
 # Setup: we optimize queries from the Join Order Benchmark on a Postgres database
-postgres_db = db.postgres.connect()
+postgres_db = postgres.connect()
 job_workload = workloads.job()
 
 # Now let's generate the optimization pipeline with our new cardinality estimator