+++ tests/test_tonic_optimization.py
@@ -12,7 +12,7 @@ from postbound.optimizer.strategies import tonic
 from tests import regression_suite
 
 
-workloads.workloads_base_dir = "../workloads"
+workloads.workloads_base_dir = "workloads/"
 pg_connect_dir = "."