+++ tests/test_ues_optimization.py
@@ -21,7 +21,7 @@ from postbound.optimizer.strategies import ues
 
 from tests import regression_suite
 
-workloads.workloads_base_dir = "../workloads"
+workloads.workloads_base_dir = "workloads/"
 pg_connect_dir = "."